pyproject.toml
[project]
name = "spec2ir-demo"
version = "0.1.0"
description = "Spec (YAML) -> Test IR (YAML) demo using pluggable LLM + Playwright a11y tree"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0.1",
  "pydantic>=2.6.0",
  "httpx>=0.27.0",
  "playwright>=1.41.0"
]

[project.scripts]
spec2ir = "spec2ir.main:main"
